+++ components/apps/src/dispatch.rs
@@ -19,16 +19,26 @@ use trussed_auth::{AuthBackend, AuthContext, AuthExtension, MAX_HW_KEY_LEN};
 #[cfg(feature = "backend-rsa")]
 use trussed_rsa_alloc::SoftwareRsa;
 
+#[cfg(feature = "backend-staging")]
+use trussed_staging::{
+    streaming::ChunkedExtension, wrap_key_to_file::WrapKeyToFileExtension, StagingBackend,
+    StagingContext,
+};
+
 #[derive(Debug)]
 pub struct Dispatch {
     #[cfg(feature = "backend-auth")]
     auth: AuthBackend,
+    #[cfg(feature = "backend-staging")]
+    staging: StagingBackend,
 }
 
-#[derive(Debug, Default)]
+#[derive(Default)]
 pub struct DispatchContext {
     #[cfg(feature = "backend-auth")]
     auth: AuthContext,
+    #[cfg(feature = "backend-staging")]
+    staging: StagingContext,
 }
 
 impl Dispatch {
@@ -38,12 +48,17 @@ impl Dispatch {
         Self {
             #[cfg(feature = "backend-auth")]
             auth: AuthBackend::new(auth_location),
+            #[cfg(feature = "backend-staging")]
+            staging: StagingBackend::new(),
         }
     }
+
     #[cfg(feature = "backend-auth")]
     pub fn with_hw_key(auth_location: Location, hw_key: Bytes<MAX_HW_KEY_LEN>) -> Self {
         Self {
             auth: AuthBackend::with_hw_key(auth_location, hw_key),
+            #[cfg(feature = "backend-staging")]
+            staging: StagingBackend::new(),
         }
     }
 }
@@ -68,6 +83,11 @@ impl ExtensionDispatch for Dispatch {
             }
             #[cfg(feature = "backend-rsa")]
             Backend::SoftwareRsa => SoftwareRsa.request(&mut ctx.core, &mut (), request, resources),
+            #[cfg(feature = "backend-staging")]
+            Backend::Staging => {
+                self.staging
+                    .request(&mut ctx.core, &mut ctx.backends.staging, request, resources)
+            }
         }
     }
 
@@ -88,9 +108,30 @@ impl ExtensionDispatch for Dispatch {
                     request,
                     resources,
                 ),
+                #[allow(unreachable_patterns)]
+                _ => Err(TrussedError::RequestNotAvailable),
             },
             #[cfg(feature = "backend-rsa")]
             Backend::SoftwareRsa => Err(TrussedError::RequestNotAvailable),
+            #[cfg(feature = "backend-staging")]
+            Backend::Staging => match extension {
+                Extension::Chunked => <StagingBackend as ExtensionImpl<ChunkedExtension>>::extension_request_serialized(
+                    &mut self.staging,
+                    &mut ctx.core,
+                    &mut ctx.backends.staging,
+                    request,
+                    resources,
+                ),
+                Extension::WrapKeyToFile => <StagingBackend as ExtensionImpl<WrapKeyToFileExtension>>::extension_request_serialized(
+                    &mut self.staging,
+                    &mut ctx.core,
+                    &mut ctx.backends.staging,
+                    request,
+                    resources,
+                ),
+                #[allow(unreachable_patterns)]
+                _ => Err(TrussedError::RequestNotAvailable),
+            },
         }
     }
 }
@@ -101,12 +142,18 @@ pub enum Backend {
     Auth,
     #[cfg(feature = "backend-rsa")]
     SoftwareRsa,
+    #[cfg(feature = "backend-staging")]
+    Staging,
 }
 
 #[derive(Debug, Clone, Copy)]
 pub enum Extension {
     #[cfg(feature = "backend-auth")]
     Auth,
+    #[cfg(feature = "backend-staging")]
+    Chunked,
+    #[cfg(feature = "backend-staging")]
+    WrapKeyToFile,
 }
 
 impl From<Extension> for u8 {
@@ -114,6 +161,10 @@ impl From<Extension> for u8 {
         match extension {
             #[cfg(feature = "backend-auth")]
             Extension::Auth => 0,
+            #[cfg(feature = "backend-staging")]
+            Extension::Chunked => 1,
+            #[cfg(feature = "backend-staging")]
+            Extension::WrapKeyToFile => 2,
         }
     }
 }
@@ -125,6 +176,10 @@ impl TryFrom<u8> for Extension {
         match id {
             #[cfg(feature = "backend-auth")]
             0 => Ok(Extension::Auth),
+            #[cfg(feature = "backend-staging")]
+            1 => Ok(Extension::Chunked),
+            #[cfg(feature = "backend-staging")]
+            2 => Ok(Extension::WrapKeyToFile),
             _ => Err(TrussedError::InternalError),
         }
     }
@@ -135,4 +190,18 @@ impl ExtensionId<AuthExtension> for Dispatch {
     type Id = Extension;
 
     const ID: Self::Id = Self::Id::Auth;
+}
+
+#[cfg(feature = "backend-staging")]
+impl ExtensionId<ChunkedExtension> for Dispatch {
+    type Id = Extension;
+
+    const ID: Self::Id = Self::Id::Chunked;
+}
+
+#[cfg(feature = "backend-staging")]
+impl ExtensionId<WrapKeyToFileExtension> for Dispatch {
+    type Id = Extension;
+
+    const ID: Self::Id = Self::Id::WrapKeyToFile;
 }